hiding character works
