+++ core/src/ecs/systems/ui/HealthBarSystem2.ts
@@ -1,26 +1,28 @@
 import { Entity, Health, HealthBar, Position, Renderable, ClientSystemBuilder, entries } from "@piggo-gg/core";
 
-// HealthBarSystem displays character health bars
+// HealthBarSystem displays gui elements
 export const HealthBarSystem = ClientSystemBuilder({
   id: "HealthBarSystem",
   init: ({ world }) => {
     if (!world.renderer) return undefined;
 
+    // let renderedEntities: Set<Entity> = new Set();
     const characterHealthBars: Record<string, Entity<Renderable | Position>> = {};
 
-    const healthbarForEntity = (entity: Entity<Health | Position | Renderable>) => {
-      if (entity.components.renderable) {
-        const { health, position } = entity.components;
+    const healthbarForCharacter = (character: Entity<Health | Position | Renderable>) => {
+      if (character.components.renderable) {
+        const { health, position } = character.components;
 
         const healthbar = Entity<Position | Renderable>({
-          id: `${entity.id}-health`,
+          id: `${character.id}-health`,
           components: {
             position: position,
             renderable: HealthBar({ health })
           }
         });
 
-        characterHealthBars[entity.id] = healthbar;
+        // renderedEntities.add(entity);
+        characterHealthBars[character.id] = healthbar;
         world.addEntity(healthbar);
       }
     }
@@ -29,27 +31,28 @@ export const HealthBarSystem = ClientSystemBuilder({
       id: "HealthBarSystem",
       query: ["health", "position", "renderable"],
       skipOnRollback: true,
-      onTick: (entities: Entity<Health | Position | Renderable>[]) => {
+      onTick: (characters: Entity<Health | Position | Renderable>[]) => {
         // handle old entities
+        // renderedEntities.forEach((entity) => {
         entries(characterHealthBars).forEach(([entityId, healthbar]) => {
           if (!world.entities[entityId]) {
-            console.log("DELETEDDD")
             world.removeEntity(healthbar.id);
+            // renderedEntities.delete(entity);
             delete characterHealthBars[entityId];
           }
         });
-  
+
         // handle new entities
-        entities.forEach((entity) => {
-          const { health, position, renderable } = entity.components;
+        characters.forEach((character) => {
+          const { health, position, renderable } = character.components;
           if (health && health.showHealthBar && position) {
-            if (!characterHealthBars[entity.id] || position !== characterHealthBars[entity.id].components.position) {
-              healthbarForEntity(entity);
+            // if (!renderedEntities.has(entity)) {
+            if (!characterHealthBars[character.id]) {
+              healthbarForCharacter(character);
             }
           }
 
-          // update visibility
-          characterHealthBars[entity.id].components.renderable.visible = renderable.visible;
+          characterHealthBars[character.id].components.renderable.visible = renderable.visible;
         });
       }
     }
